## Contributing Guidelines

<!-- Please make sure you have read and understood our contributing guidelines before submitting this PR -->
- [x] I have read and understood the [CONTRIBUTING.md](https://github.com/Autodesk/XLB/blob/main/CONTRIBUTING.md) guidelines


## Description

The mesh_boundary_masker previously used warp aabb intersection testing to determine when voxels intersect with the mesh. This just found the voxels that intersected with the triangle bounding boxes and not the triangles themselves. I added warp functions to do the triangle intersection testing after the aabb test was passed, so only those voxels that actually intersect the triangles are made solid.

<!-- 
Thank you for your contribution! Please include a summary of the changes and the related issue. Please also include relevant motivation and context. List any dependencies that are required for this change.
-->


## Type of change

<!-- Please select the options that are relevant. -->
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] Documentation update

## How Has This Been Tested?

<!-- Please describe the tests that you ran to verify your changes. Provide instructions so others can reproduce. Include details of your test environment, and the test cases you ran.

- [ ] Test A
- [ ] Test B
-->
- [ ] All pytest tests pass

All tests passing except JAX macroscopic (will be fixed in other PR)

<!-- To run the tests, execute the following command from the root of the repository:

```bash
pytest
```
 -->


## Linting and Code Formatting

Make sure the code follows the project's linting and formatting standards. This project uses **Ruff** for linting.

To run Ruff, execute the following command from the root of the repository:

```bash
ruff check .
```

<!-- You can also fix some linting errors automatically using Ruff:

```bash
ruff check . --fix
```
-->

- [x] Ruff passes

Old:
![windtunnel_3d_1000_old](https://github.com/user-attachments/assets/e9be0e8a-31c6-4756-a9ba-e29b3f53a4d9)

New:
![windtunnel_3d_1000_new](https://github.com/user-attachments/assets/b9629e3f-e4f6-4c76-a56a-1041eb6bf7b6)

Timing for SAE_Coarse stl file and the cached code:
New mesh masker completed in 0.004960 seconds.
Old mesh masker completed in 0.004057 seconds.

For DrivAer-Notchback.stl:
New method: 0.45 sec
Old method :  0.457865 seconds
Older method (wp.mesh_query_point_no_sign): 0.457386