@mehdiataei 2025-01-09
Thanks Nigel. A couple of things:
1. It doesn't look like ruff formatting is applied. I see extra spaces, spaces missing, etc. that should be easily fixed by ruff formatter.
2. Can you improve the naming of the variables? We usually use more descriptive longer names that would help with others understanding the code. 

Thanks.