@BIackthorn 2025-01-09
You can do this yourself with a Squash merge: