@mehdiataei 2025-01-10
I can. But it can lead to bad commit messages. If you couldn't do it I'll squash and merge. A few more comments:

1. ne0, ne1, and de in mesh_voxel_intersect are not immediately intuitive. Can you use better names for them? such as `normal_edge0_pre` etc.?
2. Can you add detailed docstrings explaining the purpose, inputs, outputs, and any assumptions for each function?
3. Currently, there's no check for triangles with zero normals. For example, with zero normal both dist1 and dist2 become zero and if (0 + 0) * (0 + 0) <= 0.0:  will pass and falsely indicating an intersection with every voxel box.