@mehdiataei 2025-01-14
Can you pls add the check for zero normal?